Install a project's pinned SDK on demand (dotnet-sdk-provisioning)
Newtonsoft.Json now indexes through the docker runtime: 0 files → 945
files / 22,678 symbols. It failed on a NESTED Src/global.json pinning SDK
9.0.300 while kenn provisions from the root pin (none → SDK 10), and the
Roslyn BuildHost honored the nested pin: "no compatible .NET SDK".

Mechanism (per the researched design): the sidecar TRIGGERS, kenn-toolchain
INSTALLS. On the specific hostfxr SDK-resolution failure, and only with
--provision-sdk on, kenn-dotnet reads the project's effective global.json
and shells out to a new `kenn-toolchain provision-sdk <version> <roll>
<root>`. That reuses the entrypoint's existing download+SHA-512-verify code
and installs INTO the active DOTNET_ROOT — one root, many sdk/<v>/, how
.NET does side-by-side — so the BuildHost finds it with no re-registration.
Atomic: runtime/host pieces merged additively, sdk/<version> renamed in
last as the completion marker.

Off by default: it reaches the network at index time, and off keeps an
unsatisfiable pin the named, terminal failure it is today. Verified both
ways — flag on installs SDK 9.0.316 beside 10.0.302 and indexes; flag off
installs nothing and stays at 0.

Two fixes fell out of reaching this far into a real large repo:
- MSBuildWorkspace loads with EnableWindowsTargeting, so net45/netstandard
  projects resolve their reference assemblies on Linux (restore already
  passed it; the load did not).
- nargs/targs widened u8 → u16. A 257-arg method in Newtonsoft overflowed
  the byte and failed the whole index at parse time. col_u8 → col_u16
  (it truncated on read too); the DB column is already INTEGER, no reindex.

Every new unit is tested; provision_sdk end-to-end with injected fetch/
install, the atomic merge, the digest-algorithm pick, the u16 parse, and
the flag default. Verified against real Newtonsoft.Json through docker.

+++ crates/kenn-model/src/record.rs
@@ -54,10 +54,14 @@ pub struct SymbolRecord {
     pub enclosing_sym_id: ShortId,
     #[serde(default)]
     pub partial: bool,
+    // u16, not u8: real generated/large-signature code exceeds 255. A method in
+    // Newtonsoft.Json overflowed a u8 here (value 257) and failed the whole
+    // C# index at parse time. The DB column is already INTEGER, so this is not
+    // a stored-format change.
     #[serde(default)]
-    pub nargs: u8,
+    pub nargs: u16,
     #[serde(default)]
-    pub targs: u8,
+    pub targs: u16,
     /// Denormalized from `packages[pkg].external`. `pkg = 0` → `false`.
     #[serde(default)]
     pub external: bool,
